backend/requirements.txt
transformers
torch
Pillow
python-multipart
requests
fastapi
uvicorn[standard]
authlib
splunk-opentelemetry[all]
traceloop-sdk